Detalles sugeridos por el PO para el borrado de perfiles.

+++ Opiniometro_WebApp/Opiniometro_WebApp/Controllers/PerfilController.cs
@@ -105,17 +105,30 @@ namespace Opiniometro_WebApp.Controllers
         // POST: CRUDPERFILES/Delete/5
         [HttpPost, ActionName("ConfirmarBorrado")]
         [ValidateAntiForgeryToken]
-        public ActionResult BorrarConfirmado(string id)
+        public ActionResult BorrarConfirmado(string nombre_perfil)
         {
-            Perfil perfil = db.Perfil.Find(id);
-            db.Perfil.Remove(perfil);
-            db.SaveChanges();
-            return RedirectToAction("Borrar");
+            // No se puede eliminar el perfil administrador.
+            if(nombre_perfil != "Administrador")
+            {
+                Perfil perfil = db.Perfil.Find(nombre_perfil);
+                db.Perfil.Remove(perfil);
+                db.SaveChanges();
+                return RedirectToAction("Borrar");
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.NotAcceptable);
+            }
         }
 
         public ActionResult Borrar()
         {
             return View(db.Perfil.ToList());
+        }
+        // Validación de datos.
+        public JsonResult IsNombreAvailable(string Nombre)
+        {
+            return Json(!db.Perfil.Any(item => item.Nombre == Nombre), JsonRequestBehavior.AllowGet);
         }
     }
 }
